Improved as_common_basis()
